Residuals \(\left( \boldsymbol{\hat{\varepsilon}} = \mathbf{My} \right)\)

Calculates residuals using $$ \boldsymbol{\hat{\varepsilon}} = \mathbf{My} . $$ where $$ \mathbf{M} = \mathbf{I} - \mathbf{P} \\ = \mathbf{I} - \mathbf{X} \left( \mathbf{X}^{T} \mathbf{X} \right)^{-1} \mathbf{X}^{T} . $$

My(X, y)

Arguments

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

Value

Returns an \(n \times 1\) matrix of residuals \(\left( \boldsymbol{\hat{\varepsilon}} \right)\), that is, the difference between the observed \(\left( \mathbf{y} \right)\) and predicted \(\left( \mathbf{\hat{y}} \right)\) values of the regressand variable \(\left( \boldsymbol{\hat{\varepsilon}} = \mathbf{y} - \mathbf{\hat{y}} \right)\).

References

Wikipedia: Errors and Residuals

See also

Examples

X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
My(X = X, y = y)

#>                 wages
#>    [1,]   2.032471421
#>    [2,]  -1.648829304
#>    [3,]  -5.240575302
#>    [4,] -10.862585942
#>    [5,]   7.883187386
#>    [6,]   0.068175880
#>    [7,]   1.482110593
#>    [8,]  -1.951493586
#>    [9,]  -3.018325023
#>   [10,]   2.979763500
#>   [11,]  -5.387483661
#>   [12,]  -3.001643706
#>   [13,]   0.167030965
#>   [14,]  -1.777451536
#>   [15,]   2.144652830
#>   [16,]   6.211567868
#>   [17,]   2.636511988
#>   [18,]  -9.433879255
#>   [19,]  -6.666579881
#>   [20,]   8.500748304
#>   [21,]   3.783914733
#>   [22,]   3.483553905
#>   [23,]  -4.350922050
#>   [24,]  -3.176158344
#>   [25,]  -1.613865175
#>   [26,]   5.737784758
#>   [27,]   6.307956783
#>   [28,]  13.451628576
#>   [29,]   0.147995466
#>   [30,]   4.214391287
#>   [31,]  -1.534895446
#>   [32,]  13.128581625
#>   [33,]   8.563359411
#>   [34,]  -2.028173192
#>   [35,]   2.871296331
#>   [36,]   2.003480828
#>   [37,]   4.941711082
#>   [38,]  -0.518069821
#>   [39,]   6.468642333
#>   [40,]  12.117513221
#>   [41,]  -7.708793188
#>   [42,]  -1.018611063
#>   [43,]  -1.306338758
#>   [44,]   0.130095745
#>   [45,]   0.740267770
#>   [46,]  -2.739673232
#>   [47,]  -4.354445100
#>   [48,]   1.983012663
#>   [49,]  -2.273486529
#>   [50,]  -4.759912644
#>   [51,]  -3.238288918
#>   [52,]  -0.424715032
#>   [53,]   0.478046301
#>   [54,]  -0.064375375
#>   [55,]   1.240448184
#>   [56,]  -2.714353349
#>   [57,]  -6.155887663
#>   [58,]  -1.810936130
#>   [59,]   1.780185036
#>   [60,]  -0.786330369
#>   [61,]  -3.650260875
#>   [62,]   3.334359150
#>   [63,]  -7.349839566
#>   [64,]  -0.580593120
#>   [65,]  -4.072034828
#>   [66,]  -1.304676350
#>   [67,]   2.149438778
#>   [68,]   0.900545864
#>   [69,]  -9.091141147
#>   [70,]  -2.317357182
#>   [71,]  -2.759371402
#>   [72,]  -2.389611669
#>   [73,]  -1.353723444
#>   [74,]  -6.054676350
#>   [75,]  -1.573697237
#>   [76,]   2.294059042
#>   [77,]  -3.051018864
#>   [78,]  12.275188154
#>   [79,]   5.911328229
#>   [80,]   0.336781693
#>   [81,]  -8.917121123
#>   [82,]  -4.850561222
#>   [83,]   0.331891496
#>   [84,]  -4.752456364
#>   [85,]  11.075804184
#>   [86,]  -4.760093058
#>   [87,]  -0.581314776
#>   [88,]   1.744391287
#>   [89,]   0.573480828
#>   [90,]  -1.935398006
#>   [91,]  -6.284706643
#>   [92,]  -3.509371402
#>   [93,]   9.144652830
#>   [94,]  -1.202004534
#>   [95,] -10.208083901
#>   [96,]   9.481267509
#>   [97,]  -0.752043845
#>   [98,]   3.074217419
#>   [99,]  -2.148648890
#>  [100,]  -2.760093058
#>  [101,]   6.237661726
#>  [102,]   5.897235127
#>  [103,]   0.296183058
#>  [104,]  -1.693719262
#>  [105,]   9.056142975
#>  [106,]  -4.450206085
#>  [107,]   3.517777224
#>  [108,]  11.055782147
#>  [109,]  -2.093486529
#>  [110,]  -8.263208435
#>  [111,]   3.553998561
#>  [112,]  -0.332645167
#>  [113,]   0.923929463
#>  [114,]  -1.016265681
#>  [115,]   5.499258364
#>  [116,]   5.819619192
#>  [117,] -10.437297329
#>  [118,]  -8.099839566
#>  [119,]   0.059424698
#>  [120,]   4.142702671
#>  [121,]  -2.375649000
#>  [122,]   3.147115421
#>  [123,]  -0.209371402
#>  [124,]   0.907754343
#>  [125,]   0.224841406
#>  [126,]   7.571344564
#>  [127,]   3.991365190
#>  [128,]   1.424985703
#>  [129,]   7.432651835
#>  [130,]  -3.011765123
#>  [131,]  -1.228173192
#>  [132,]  -0.027828699
#>  [133,]  -2.301132758
#>  [134,]  -5.279208806
#>  [135,]   0.158836827
#>  [136,]  -5.342584459
#>  [137,]  10.389356033
#>  [138,]  -4.925797516
#>  [139,]  -5.915339008
#>  [140,]  10.713178997
#>  [141,]   6.842998583
#>  [142,]  -5.152241449
#>  [143,]  -2.259912644
#>  [144,]   0.647995466
#>  [145,]  -1.627019235
#>  [146,]   1.103661242
#>  [147,]  -5.938565918
#>  [148,]  -3.760273472
#>  [149,]  -0.531703559
#>  [150,]  -1.387980302
#>  [151,]   4.889244284
#>  [152,]  -3.763486529
#>  [153,]  23.602035805
#>  [154,]  24.574382898
#>  [155,]  -3.236158344
#>  [156,]   4.889244284
#>  [157,]  -3.017708993
#>  [158,]  -3.082847607
#>  [159,]  -4.682667193
#>  [160,]  -3.183954694
#>  [161,]  -0.407325284
#>  [162,]   1.552380526
#>  [163,]   4.148717122
#>  [164,]   4.978624516
#>  [165,]  -1.851824120
#>  [166,]   3.495678788
#>  [167,]   7.305382648
#>  [168,]   3.699392800
#>  [169,]  20.341433855
#>  [170,]  -8.675261965
#>  [171,]  -4.933238728
#>  [172,]  -1.194779721
#>  [173,]   5.074382898
#>  [174,]   5.074382898
#>  [175,]  -3.211134362
#>  [176,]  -0.217528579
#>  [177,]  -3.333788360
#>  [178,]   2.093337386
#>  [179,]   0.238897536
#>  [180,]   9.288717122
#>  [181,]  -2.469093308
#>  [182,]   4.220476995
#>  [183,]   3.515504064
#>  [184,]   4.013480828
#>  [185,]  -1.498191238
#>  [186,]  -1.143774280
#>  [187,]   4.357776369
#>  [188,]  12.408295585
#>  [189,]  -4.384931552
#>  [190,]   0.275670398
#>  [191,]  -2.185037178
#>  [192,]   5.200726278
#>  [193,]  -9.090780319
#>  [194,]   5.728046301
#>  [195,]  -0.852004534
#>  [196,]  -7.929659152
#>  [197,]  -4.833208435
#>  [198,]  -5.426699587
#>  [199,]  -3.483413452
#>  [200,]   3.491350254
#>  [201,]   3.845323650
#>  [202,]   9.015684478
#>  [203,]  -5.491282878
#>  [204,]   4.759642072
#>  [205,]   2.108656413
#>  [206,]   8.145735314
#>  [207,]   3.328115171
#>  [208,]  -2.279930462
#>  [209,]   1.146572468
#>  [210,]  -8.573088729
#>  [211,]   1.240267770
#>  [212,]   6.308175880
#>  [213,]  -0.009551816
#>  [214,]   4.644833244
#>  [215,]  -1.741296958
#>  [216,]  -1.509912644
#>  [217,]  -5.020394888
#>  [218,]   0.717934757
#>  [219,]  -0.185398006
#>  [220,]  -1.426519172
#>  [221,]   5.395670398
#>  [222,]  -4.694238479
#>  [223,]   0.882441128
#>  [224,]   5.399131257
#>  [225,]   1.953720467
#>  [226,]  -4.838565918
#>  [227,]  -1.327992778
#>  [228,]   6.614391287
#>  [229,]  -1.947469354
#>  [230,]   2.357499141
#>  [231,]  -5.166421492
#>  [232,]   7.520628598
#>  [233,]  -0.307708993
#>  [234,]  -3.969478738
#>  [235,]   3.018679294
#>  [236,]   2.073661242
#>  [237,]  -2.129371402
#>  [238,]  -2.047469582
#>  [239,]  -4.517528579
#>  [240,]   7.572230299
#>  [241,]  -2.027812364
#>  [242,]  -0.259912644
#>  [243,]  -2.025647396
#>  [244,]   9.074382898
#>  [245,]  -4.334088479
#>  [246,]  -6.925617102
#>  [247,]   3.109883206
#>  [248,]   3.065021820
#>  [249,]  -4.667301537
#>  [250,]  -0.482764873
#>  [251,]   1.311965657
#>  [252,]  -6.345316983
#>  [253,]  -5.543306115
#>  [254,]  -2.425797516
#>  [255,]   0.698927950
#>  [256,]  10.589785526
#>  [257,]  -0.481703559
#>  [258,]  -0.925977930
#>  [259,]  -2.603052624
#>  [260,]  -3.466446095
#>  [261,]   1.880456573
#>  [262,]   0.196702274
#>  [263,]  -3.194779721
#>  [264,]   4.844962822
#>  [265,]   0.998234877
#>  [266,] -16.805580130
#>  [267,]   4.008762039
#>  [268,]  -0.981681534
#>  [269,]   0.546882688
#>  [270,]  -6.806301786
#>  [271,]   8.074360872
#>  [272,]  -3.100741636
#>  [273,]   2.095954411
#>  [274,]  -4.425075860
#>  [275,]  -4.747438311
#>  [276,]   6.315504064
#>  [277,]   1.664202484
#>  [278,]  -0.921463292
#>  [279,]  -1.212215242
#>  [280,]   1.166813591
#>  [281,]  -4.229719005
#>  [282,]  -7.332284339
#>  [283,]   1.062110593
#>  [284,]   4.243698214
#>  [285,]   0.850809868
#>  [286,]   2.365058781
#>  [287,]   6.031289546
#>  [288,]  -5.593427532
#>  [289,]   2.177679555
#>  [290,]   8.760817401
#>  [291,]  -4.311141147
#>  [292,]  -5.683052624
#>  [293,]  -2.307181831
#>  [294,]   7.908836827
#>  [295,]  -6.267353856
#>  [296,]  -1.470214247
#>  [297,]   4.978326000
#>  [298,]   5.880335146
#>  [299,]  -1.564172935
#>  [300,]   2.184383753
#>  [301,]   0.593035544
#>  [302,]   6.009419007
#>  [303,]   2.549363989
#>  [304,]  -8.923834988
#>  [305,]  -3.707631950
#>  [306,]   3.240628598
#>  [307,]  -4.528853668
#>  [308,]   5.834201000
#>  [309,]  -1.869543427
#>  [310,]   5.437665052
#>  [311,]  -2.361566663
#>  [312,]   3.901409479
#>  [313,]  -0.734949370
#>  [314,]   1.973270120
#>  [315,] -12.800575302
#>  [316,]  -2.242186659
#>  [317,]  -3.183954694
#>  [318,]   7.745284968
#>  [319,]  -0.112945287
#>  [320,]  -2.444960135
#>  [321,]   0.967354833
#>  [322,]  -3.600110876
#>  [323,] -12.156821003
#>  [324,]  -6.849298324
#>  [325,]  -1.572602277
#>  [326,]   1.114601994
#>  [327,]  -5.166421492
#>  [328,]  -2.593125701
#>  [329,] -10.187116915
#>  [330,]   0.546702274
#>  [331,]   3.880456573
#>  [332,]  15.687053002
#>  [333,]  -0.626421492
#>  [334,]   7.244962822
#>  [335,]   4.518101091
#>  [336,]  -1.597348165
#>  [337,]  -1.508108504
#>  [338,]  -6.176806923
#>  [339,]   3.002187636
#>  [340,]   0.117693863
#>  [341,]  -0.566001678
#>  [342,]  -1.907203856
#>  [343,]   2.833939336
#>  [344,]  10.645194072
#>  [345,]  -3.540690800
#>  [346,]  11.476142975
#>  [347,]  -4.961349277
#>  [348,]  16.927294124
#>  [349,]  -2.991959617
#>  [350,]  -5.308191238
#>  [351,]   0.483012663
#>  [352,]   3.714721700
#>  [353,]   6.904952950
#>  [354,]  -1.055857369
#>  [355,]   5.392471421
#>  [356,]  -3.834088479
#>  [357,]   4.919063870
#>  [358,]  -1.189154016
#>  [359,]  -1.085037178
#>  [360,]  -1.123968774
#>  [361,]  -4.064986342
#>  [362,]   0.993939336
#>  [363,]  11.214322189
#>  [364,]   7.147874038
#>  [365,]  -6.855707998
#>  [366,]   1.012832249
#>  [367,]   0.092276927
#>  [368,]  13.572035805
#>  [369,]   0.240628598
#>  [370,]   9.399258364
#>  [371,]  26.933193077
#>  [372,]  -2.435217592
#>  [373,]  -8.436024559
#>  [374,]   0.440636987
#>  [375,]  -1.602004534
#>  [376,]  -4.861205835
#>  [377,]   3.853854880
#>  [378,]   7.005670398
#>  [379,]  -3.238251946
#>  [380,]   5.232110593
#>  [381,]  -4.017708993
#>  [382,]  -0.723005140
#>  [383,]   2.315143236
#>  [384,]   4.779063870
#>  [385,]  -3.644135108
#>  [386,]  -0.518250235
#>  [387,]  -3.944779721
#>  [388,]  -0.113323933
#>  [389,]  -4.481320706
#>  [390,]   2.978767958
#>  [391,]  -1.259912644
#>  [392,]   0.408295585
#>  [393,]  -1.642796771
#>  [394,]   7.013720239
#>  [395,]   2.417513221
#>  [396,]  -3.342584459
#>  [397,]  -1.926482200
#>  [398,]   6.621379065
#>  [399,]  -8.409190988
#>  [400,]   8.640146354
#>  [401,]  -8.349478738
#>  [402,]  -0.240755716
#>  [403,]  -4.473413452
#>  [404,]   0.933553905
#>  [405,]  -1.505316127
#>  [406,]  -2.768430649
#>  [407,]  -3.100741636
#>  [408,]  -8.984361500
#>  [409,] -10.072727901
#>  [410,]  -4.500514594
#>  [411,]  -2.307528579
#>  [412,]  -4.405436688
#>  [413,]  -2.684856764
#>  [414,]  -4.908852185
#>  [415,]  -3.926699587
#>  [416,]  -2.016987337
#>  [417,]  -4.718430649
#>  [418,]  -1.685217592
#>  [419,]   6.184878367
#>  [420,]  -4.593486529
#>  [421,]  -2.093306115
#>  [422,]  -4.602404045
#>  [423,]   2.720069538
#>  [424,]   1.154180458
#>  [425,]   3.491470815
#>  [426,]  -1.074329602
#>  [427,]  -1.257408873
#>  [428,]  -4.254599307
#>  [429,]   5.645112943
#>  [430,]  -0.926699587
#>  [431,]   1.496947376
#>  [432,]  -1.280936130
#>  [433,]   6.058942453
#>  [434,]   3.741169840
#>  [435,]  -2.591884829
#>  [436,]  -5.916241078
#>  [437,]   1.573661242
#>  [438,]   4.226392054
#>  [439,] -13.240214474
#>  [440,]  -4.860093058
#>  [441,]  -3.270510386
#>  [442,]   1.147604344
#>  [443,] -10.404173790
#>  [444,]  -3.884495936
#>  [445,]   0.978767958
#>  [446,]  -3.056579025
#>  [447,]  -4.010093058
#>  [448,]  -1.785247885
#>  [449,]   1.730817401
#>  [450,]   4.536633177
#>  [451,]   8.769544403
#>  [452,]  -3.384495936
#>  [453,]  16.571869471
#>  [454,]  -4.816866776
#>  [455,]  -8.884895446
#>  [456,]   1.353088102
#>  [457,]   1.316766962
#>  [458,]   6.610069538
#>  [459,]   0.583939336
#>  [460,]  -0.818069821
#>  [461,]   0.647995466
#>  [462,]  -2.851463292
#>  [463,]   4.483553905
#>  [464,]  -1.855707998
#>  [465,]  17.822909292
#>  [466,]   1.500185036
#>  [467,]  -6.499273722
#>  [468,]   5.058235105
#>  [469,]   0.713489217
#>  [470,]  -1.356249240
#>  [471,]   0.572368050
#>  [472,]   4.636663591
#>  [473,]  11.515504064
#>  [474,]  -5.016265681
#>  [475,]   2.112651835
#>  [476,]  -8.999814964
#>  [477,]  -4.834315522
#>  [478,]  -2.367560477
#>  [479,]  -7.249634550
#>  [480,]  12.132291007
#>  [481,]   0.905465382
#>  [482,]  -1.321608456
#>  [483,]  -0.713954694
#>  [484,]  -1.633774280
#>  [485,]  -0.062764873
#>  [486,]   0.483553905
#>  [487,]  -7.666241078
#>  [488,]  -3.091862803
#>  [489,]   0.985864892
#>  [490,]  -0.759912644
#>  [491,]   5.520809012
#>  [492,]  13.097513221
#>  [493,]  -3.016265681
#>  [494,]   7.943900653
#>  [495,]   0.593035544
#>  [496,]   1.844689802
#>  [497,]   8.130726278
#>  [498,]   2.445013658
#>  [499,]  -0.131825831
#>  [500,]  -1.325333945
#>  [501,]  -6.676821003
#>  [502,]  -0.258130529
#>  [503,]   2.806615359
#>  [504,]  -2.276880001
#>  [505,]  21.043878628
#>  [506,]   8.190485156
#>  [507,]  -0.445580130
#>  [508,]  18.633775257
#>  [509,] -11.766903737
#>  [510,]  -2.114406417
#>  [511,]  18.553720239
#>  [512,]  12.224119750
#>  [513,]  27.007474538
#>  [514,]  -2.759190988
#>  [515,]   2.267693635
#>  [516,]  -1.813834133
#>  [517,]   0.814962822
#>  [518,]   7.566406134
#>  [519,]  -3.942270259
#>  [520,]  -2.332125951
#>  [521,]  -0.595227464
#>  [522,]  -3.001282878
#>  [523,]  20.093359411
#>  [524,]  -2.188018985
#>  [525,]  -1.676446095
#>  [526,]  -2.350200394
#>  [527,]  -2.259912644
#>  [528,]  -4.332847607
#>  [529,]  -6.350566913
#>  [530,]  -1.517167751
#>  [531,]  -0.814714177
#>  [532,]  -4.943398721
#>  [533,]  -3.517889407
#>  [534,] -10.091501975
#>  [535,]   5.970837823
#>  [536,]  -1.063323933
#>  [537,]   4.990989426
#>  [538,]  -2.085217592
#>  [539,]  -2.935037178
#>  [540,]  -5.847838571
#>  [541,]   0.375323650
#>  [542,]  14.740087356
#>  [543,]  -3.850380808
#>  [544,]   1.448285713
#>  [545,] -10.441501975
#>  [546,]  -5.517301537
#>  [547,]  -2.126626509
#>  [548,]   0.647722446
#>  [549,]  -1.462223631
#>  [550,]  -7.066940709
#>  [551,]   1.948897536
#>  [552,]   0.084827326
#>  [553,]  -5.100019980
#>  [554,]  -0.361386249
#>  [555,]   1.425323650
#>  [556,]  -3.934676350
#>  [557,]  -1.500291290
#>  [558,]   0.469138658
#>  [559,]  -5.749659152
#>  [560,]  -5.187860597
#>  [561,]  -7.739695258
#>  [562,]  -3.056579025
#>  [563,]   3.075104554
#>  [564,]   3.811800601
#>  [565,]   7.645684478
#>  [566,]   6.066933296
#>  [567,]  -3.433052624
#>  [568,]  -0.319673232
#>  [569,]  -4.353968774
#>  [570,]   4.585013658
#>  [571,]  -7.308283855
#>  [572,]  -1.258830160
#>  [573,]  19.886572468
#>  [574,]  -4.026279533
#>  [575,]   4.636707337
#>  [576,]  -2.004811618
#>  [577,]  -7.740034060
#>  [578,]  11.008678439
#>  [579,]   5.714210873
#>  [580,]   0.408317611
#>  [581,]  -0.786330369
#>  [582,] -10.937303020
#>  [583,]   9.381719472
#>  [584,]  -1.351282878
#>  [585,]  -2.798144609
#>  [586,]   0.662291007
#>  [587,]  -3.849478738
#>  [588,]  17.972818169
#>  [589,]   1.547415541
#>  [590,]   2.803991776
#>  [591,]  -6.735617102
#>  [592,]   8.563359411
#>  [593,]   0.154004252
#>  [594,]   1.782110593
#>  [595,]  -1.926699587
#>  [596,]  -4.979010574
#>  [597,]  -6.833208435
#>  [598,]  12.339785526
#>  [599,]  -5.687658157
#>  [600,]  -2.963690679
#>  [601,]  -1.759551816
#>  [602,]  -0.729659152
#>  [603,]   2.437874049
#>  [604,]   0.645172047
#>  [605,]  -2.785608713
#>  [606,]  13.554022070
#>  [607,]  -1.285969541
#>  [608,]  12.106747192
#>  [609,]   0.780276159
#>  [610,]   8.274022070
#>  [611,]   1.278438172
#>  [612,]  -1.185037178
#>  [613,]   0.778897536
#>  [614,] -12.112140031
#>  [615,]  -3.035020843
#>  [616,]   6.909906942
#>  [617,]   1.239906942
#>  [618,]   2.080276159
#>  [619,]   1.837595955
#>  [620,]   8.538415291
#>  [621,]   8.640395865
#>  [622,]  -2.785789127
#>  [623,]  -9.525466982
#>  [624,]   4.006031226
#>  [625,]   0.053428509
#>  [626,]  -2.258469332
#>  [627,]  -0.645338781
#>  [628,]  -1.851102464
#>  [629,]  -1.086601906
#>  [630,]  -2.113774280
#>  [631,]  -0.671463292
#>  [632,]   1.258717122
#>  [633,]  -0.499454136
#>  [634,]  -1.035789127
#>  [635,] -10.757567262
#>  [636,]   3.148175880
#>  [637,]   0.041223147
#>  [638,]  20.719605112
#>  [639,] -10.574329602
#>  [640,]  -5.884390310
#>  [641,]  -2.221742242
#>  [642,]  -0.240214474
#>  [643,]  -0.535217592
#>  [644,]   0.506962107
#>  [645,]  -4.675256274
#>  [646,]  -6.011862803
#>  [647,]  -6.016987337
#>  [648,]   0.140809012
#>  [649,]  -5.183954694
#>  [650,]  -6.672043217
#>  [651,]  -3.843413452
#>  [652,]  -1.971102464
#>  [653,]  -2.334495936
#>  [654,]  -0.767708993
#>  [655,]  -3.699251696
#>  [656,]   8.889438778
#>  [657,]   3.741508643
#>  [658,]  -5.999454136
#>  [659,]  -4.093306115
#>  [660,]  -7.198912894
#>  [661,]  -1.685217592
#>  [662,]   6.017874049
#>  [663,]   0.583398094
#>  [664,]  -0.954856764
#>  [665,] -11.273488240
#>  [666,]  -6.350200394
#>  [667,]   5.774924140
#>  [668,]   2.729128786
#>  [669,] -10.008083901
#>  [670,]  -1.306338758
#>  [671,]  -7.349432110
#>  [672,]  -0.912245657
#>  [673,]  -5.259912644
#>  [674,]  -8.591141147
#>  [675,]   0.859394404
#>  [676,]  -2.525037178
#>  [677,]  -6.805496541
#>  [678,]  -3.855790732
#>  [679,]   3.265941935
#>  [680,]  -1.859612524
#>  [681,]   4.324053340
#>  [682,] -10.924359895
#>  [683,]   2.943850045
#>  [684,]  -1.287658157
#>  [685,]   5.315143236
#>  [686,]  -0.607477743
#>  [687,]  -5.403593866
#>  [688,]   2.800989426
#>  [689,]   4.705684478
#>  [690,]   3.358498025
#>  [691,]  -4.093306115
#>  [692,]  -5.017348165
#>  [693,]   0.121569351
#>  [694,]  -5.679010574
#>  [695,]  -6.426158344
#>  [696,]  -6.041736312
#>  [697,]  15.909219681
#>  [698,]   0.425314633
#>  [699,]  -6.591884829
#>  [700,]  -5.113684761
#>  [701,]   0.477865887
#>  [702,]   2.723253785
#>  [703,]  16.812555021
#>  [704,]   3.203698441
#>  [705,]   0.716807900
#>  [706,]   6.010430366
#>  [707,]  33.005309570
#>  [708,]  -0.166460175
#>  [709,]  -1.425436688
#>  [710,]   2.969948122
#>  [711,]  -0.592404045
#>  [712,]   4.855443356
#>  [713,]  -3.459839566
#>  [714,]  -1.187477743
#>  [715,]   1.590911755
#>  [716,]  50.414119750
#>  [717,]   0.907776369
#>  [718,]  -1.500334180
#>  [719,]   5.231803061
#>  [720,]  10.459424698
#>  [721,]  -2.092945287
#>  [722,]  -2.758649746
#>  [723,] -12.605803207
#>  [724,]   0.763489217
#>  [725,]   4.316225720
#>  [726,]   5.228407130
#>  [727,]  -7.487001417
#>  [728,]  -4.291682389
#>  [729,]  -2.734149188
#>  [730,]   1.886315239
#>  [731,]  -2.776158344
#>  [732,]  11.071476745
#>  [733,]   1.649258364
#>  [734,]  -0.052764873
#>  [735,]   4.458883456
#>  [736,]  37.167152393
#>  [737,]  -4.353723444
#>  [738,]  -6.100561222
#>  [739,] -10.153543030
#>  [740,]   7.088897536
#>  [741,]   9.965735314
#>  [742,]   2.316586548
#>  [743,]  -2.520171584
#>  [744,]  11.773914733
#>  [745,]  -2.942764873
#>  [746,]   2.310898531
#>  [747,]  -1.925256274
#>  [748,]  -5.072796771
#>  [749,]   8.762577047
#>  [750,]  -6.324149188
#>  [751,]  -3.140455597
#>  [752,]  -1.349478738
#>  [753,]   7.479670028
#>  [754,]  -5.108310943
#>  [755,]   1.378557250
#>  [756,] -12.507933780
#>  [757,]  -4.350380808
#>  [758,]   3.098822747
#>  [759,]   0.073480828
#>  [760,]  -1.355888412
#>  [761,]   0.388613751
#>  [762,]   4.610576279
#>  [763,]  17.828942453
#>  [764,]   0.880095745
#>  [765,]  -1.934676350
#>  [766,]  -0.444435227
#>  [767,]   0.572252324
#>  [768,]  -3.259912644
#>  [769,] -20.780516305
#>  [770,]   0.018296441
#>  [771,]  -5.166038639
#>  [772,]  -2.980845007
#>  [773,]   0.542998811
#>  [774,]   0.258461919
#>  [775,]  -4.158469332
#>  [776,]  -1.896158344
#>  [777,]  -3.838560227
#>  [778,]  -3.693877651
#>  [779,]  17.434924140
#>  [780,]  -0.445995748
#>  [781,]  15.517332807
#>  [782,]   0.315143236
#>  [783,]  -5.156911898
#>  [784,]  -5.222463898
#>  [785,]  -2.267528579
#>  [786,]   3.501087106
#>  [787,]  -4.355527584
#>  [788,]  -4.022314527
#>  [789,]   2.334119750
#>  [790,]  -4.275617102
#>  [791,]  -1.364264686
#>  [792,]  -7.600200394
#>  [793,]  -7.907903487
#>  [794,]   2.814601994
#>  [795,]  -4.861386249
#>  [796,]  13.710868237
#>  [797,]   8.907054713
#>  [798,]   2.285382875
#>  [799,]   5.147585660
#>  [800,]  27.927091685
#>  [801,]   1.759620047
#>  [802,]   3.502374607
#>  [803,]   1.321344564
#>  [804,]  -7.949432110
#>  [805,]   4.979424698
#>  [806,]  -6.447631950
#>  [807,]  22.582216219
#>  [808,]  24.553102182
#>  [809,]  30.886572468
#>  [810,]   2.708317611
#>  [811,]   4.666406134
#>  [812,]  -6.685940958
#>  [813,]  -2.646658407
#>  [814,]  -3.684495936
#>  [815,]   3.565143236
#>  [816,]   8.136548949
#>  [817,]   7.605248862
#>  [818,]  -0.535969541
#>  [819,]  33.081267520
#>  [820,]   2.146031226
#>  [821,]   3.491147815
#>  [822,]  -9.813247118
#>  [823,]  -3.592245657
#>  [824,]  -5.066601906
#>  [825,]   0.365804184
#>  [826,]   1.877513221
#>  [827,]   1.736150932
#>  [828,]   4.770448184
#>  [829,]   4.210868237
#>  [830,]  -0.482886062
#>  [831,]  -4.916405157
#>  [832,]   5.243553905
#>  [833,]   3.811800601
#>  [834,]  -1.092584459
#>  [835,]  -3.354445100
#>  [836,]  -5.038469332
#>  [837,]  -5.434135108
#>  [838,]   5.127152393
#>  [839,]  -1.281373774
#>  [840,]  -1.605347170
#>  [841,]  -5.605490612
#>  [842,]  -5.688379813
#>  [843,]   0.805384359
#>  [844,]  -1.834315522
#>  [845,]  16.016586548
#>  [846,]  -2.750455597
#>  [847,]  -1.633774280
#>  [848,]   7.637235127
#>  [849,]  -5.901477372
#>  [850,]  -3.514209896
#>  [851,]  -3.144527834
#>  [852,]  -4.925436688
#>  [853,]  -2.338732480
#>  [854,]   6.266971979
#>  [855,]  -5.112645167
#>  [856,]   0.423300413
#>  [857,]  -1.984354204
#>  [858,]  -6.189659152
#>  [859,]   5.410002901
#>  [860,]  -4.102034828
#>  [861,]   0.171530668
#>  [862,]  -7.266085267
#>  [863,]   1.034660992
#>  [864,]   0.214210873
#>  [865,]  -3.861205835
#>  [866,]  -3.350922050
#>  [867,]  17.479183575
#>  [868,]  13.479363989
#>  [869,]  -3.478469332
#>  [870,]   0.405645796
#>  [871,]   7.995368568
#>  [872,]  -5.442270259
#>  [873,]  -1.250269264
#>  [874,]  -1.772945287
#>  [875,]  -2.759732230
#>  [876,]  -6.183413452
#>  [877,]  -2.351282878
#>  [878,]   1.758920416
#>  [879,]  17.422276927
#>  [880,]  -2.851463292
#>  [881,]   0.458536708
#>  [882,]   4.907754343
#>  [883,]   2.723539825
#>  [884,]  -0.975310198
#>  [885,]  14.213268516
#>  [886,]   7.092457341
#>  [887,]  10.638822747
#>  [888,]   4.983012663
#>  [889,]  -5.575819542
#>  [890,]  -2.711854414
#>  [891,]  -2.026423214
#>  [892,]  -6.014113082
#>  [893,]  -4.350019980
#>  [894,]  -0.447566406
#>  [895,]  -4.293992521
#>  [896,]  -6.231765123
#>  [897,]  -2.499273722
#>  [898,]  -0.949814964
#>  [899,]   2.492081155
#>  [900,] -11.157362245
#>  [901,]  -1.118611063
#>  [902,]   0.261388937
#>  [903,]   2.483914733
#>  [904,]  -3.265436688
#>  [905,] -10.258310943
#>  [906,]  -6.077903487
#>  [907,]   1.740249952
#>  [908,]  -5.676699587
#>  [909,]  -0.187838571
#>  [910,]   1.367805179
#>  [911,]  -6.675617102
#>  [912,]  -3.960093058
#>  [913,]  -2.102004534
#>  [914,]  -7.166085267
#>  [915,]  -4.571916726
#>  [916,]   9.686405268
#>  [917,]  -5.998132012
#>  [918,]   1.880817401
#>  [919,]  -5.833028021
#>  [920,]  -3.209454136
#>  [921,]  -6.609010574
#>  [922,]  -5.116911898
#>  [923,]  -4.317167751
#>  [924,]   6.235465382
#>  [925,]  -1.474805928
#>  [926,]  -1.392043217
#>  [927,]   8.858822747
#>  [928,]  -0.851824120
#>  [929,]  -1.090802345
#>  [930,]  -1.211252464
#>  [931,]  -3.996085267
#>  [932,]  -1.351282878
#>  [933,]  -3.092404045
#>  [934,]  -6.056579025
#>  [935,]   5.060267770
#>  [936,]  10.019883206
#>  [937,]  -0.139032599
#>  [938,]   1.323914733
#>  [939,]  -3.907903487
#>  [940,]  10.010868010
#>  [941,]  13.245609690
#>  [942,] -10.331945537
#>  [943,]   9.673720239
#>  [944,]  -2.553413452
#>  [945,]  -8.517181831
#>  [946,]  -4.843306115
#>  [947,]   0.719708710
#>  [948,]   2.257776369
#>  [949,]  -6.556121372
#>  [950,]  -0.106158344
#>  [951,]  -3.367348165
#>  [952,]  -1.352004534
#>  [953,]   0.929347882
#>  [954,]  -4.340741636
#>  [955,] -10.905918933
#>  [956,]   7.407595955
#>  [957,] -15.274354204
#>  [958,]   2.504022070
#>  [959,]  -0.883954694
#>  [960,]  -2.259912644
#>  [961,]  -3.223593866
#>  [962,]  -2.521773285
#>  [963,]  -9.091862803
#>  [964,]  14.410146354
#>  [965,]  -2.851282878
#>  [966,]   2.767415541
#>  [967,]  -0.481703559
#>  [968,]  -6.536370656
#>  [969,]  23.989965940
#>  [970,]   6.213823838
#>  [971,]   0.482832249
#>  [972,]  -1.999814964
#>  [973,]  -9.020871214
#>  [974,]  -2.871682389
#>  [975,]  -1.685217592
#>  [976,]   4.824563312
#>  [977,]  -9.591501975
#>  [978,]  -2.517889407
#>  [979,]  -3.734315522
#>  [980,]  -8.019298324
#>  [981,]   0.671350254
#>  [982,]  -1.454676350
#>  [983,]  -0.885918705
#>  [984,]  -5.259912644
#>  [985,]  -7.033757945
#>  [986,]  -0.281703559
#>  [987,]   2.231930179
#>  [988,]   0.720898531
#>  [989,]  -7.233052624
#>  [990,]  21.682877167
#>  [991,]   6.645532875
#>  [992,]  -7.296038639
#>  [993,]  -3.426338758
#>  [994,]  -2.852872210
#>  [995,] -10.814209896
#>  [996,]  -1.869814964
#>  [997,]  -5.109673232
#>  [998,]   2.329077950
#>  [999,]  -1.000334180
#> [1000,]  -2.560079833
#> [1001,]  -2.574510016
#> [1002,]  -4.584084272
#> [1003,]   0.713669631
#> [1004,]  -1.231538081
#> [1005,]  -1.592065243
#> [1006,]   1.316586548
#> [1007,]  -3.290755716
#> [1008,]   1.315864892
#> [1009,]  -8.690936130
#> [1010,]  -3.683774280
#> [1011,]   0.626933296
#> [1012,]   0.329128786
#> [1013,]   1.034202484
#> [1014,]  -1.591862803
#> [1015,]  -2.092764873
#> [1016,]  -0.134390310
#> [1017,]  -0.619723841
#> [1018,]  -1.935037178
#> [1019,]   5.777175274
#> [1020,] -13.157181831
#> [1021,]  -0.019310694
#> [1022,]   0.075104554
#> [1023,]  -2.676699587
#> [1024,]   6.644111588
#> [1025,]  11.154022070
#> [1026,]  -1.945977930
#> [1027,]  -5.833208435
#> [1028,]  -3.334135108
#> [1029,]  -3.943306115
#> [1030,]  -0.854986342
#> [1031,]   2.278356294
#> [1032,]   5.215684478
#> [1033,]  -8.582125951
#> [1034,]   1.912291007
#> [1035,]  -0.726399467
#> [1036,]  -4.093306115
#> [1037,]  -3.306579025
#> [1038,]   5.686031226
#> [1039,]  -1.769605739
#> [1040,]  -7.350206085
#> [1041,]   7.993878628
#> [1042,]   8.065143236
#> [1043,]  -3.091862803
#> [1044,]  -1.781193360
#> [1045,]   0.547423930
#> [1046,]  -1.233938359
#> [1047,]   0.842127783
#> [1048,]  -0.085789127
#> [1049,]   0.805220279
#> [1050,]   0.312702671
#> [1051,] -11.507747676
#> [1052,]  -4.826519172
#> [1053,]  -2.290019980
#> [1054,]  -1.654986342
#> [1055,]  -9.055533502
#> [1056,]   7.530596700
#> [1057,]   7.726601279
#> [1058,] -14.479914355
#> [1059,]  -3.852764873
#> [1060,]   3.309792311
#> [1061,]   1.557415541
#> [1062,]  -1.668611063
#> [1063,]  -3.018371652
#> [1064,]  -7.516446095
#> [1065,]   3.074382898
#> [1066,]  -1.083028021
#> [1067,]   1.083270975
#> [1068,]  -0.152004534
#> [1069,]  -5.017348165
#> [1070,]  -1.540440661
#> [1071,]  40.557368913
#> [1072,]  -0.252945287
#> [1073,]  11.405188154
#> [1074,]  -0.885969541
#> [1075,]  -0.311793706
#> [1076,]  -6.925617102
#> [1077,]   1.648739147
#> [1078,]   0.718296441
#> [1079,]  -8.257362245
#> [1080,]  -3.733593866
#> [1081,]   0.977776369
#> [1082,]  -6.500695008
#> [1083,]  -1.026820775
#> [1084,]  -3.399304015
#> [1085,]  -5.178130529
#> [1086,]   0.128897536
#> [1087,]  -1.957903487
#> [1088,]  -1.268430649
#> [1089,] -12.523788360
#> [1090,]  -0.965699836
#> [1091,]  15.916994005
#> [1092,]  -2.973005140
#> [1093,]  -5.333208435
#> [1094,]  -5.038973374
#> [1095,]   5.627294352
#> [1096,]  -2.652215242
#> [1097,]  -4.039543427
#> [1098,]   0.472007222
#> [1099,]  -4.657542659
#> [1100,]  -3.199930462
#> [1101,]  -3.259190988
#> [1102,]  -9.073247118
#> [1103,]  -2.559912644
#> [1104,]  -0.618280528
#> [1105,]  -1.790575075
#> [1106,]   4.811439773
#> [1107,]   0.959605339
#> [1108,]  -3.093125701
#> [1109,]  -1.043924279
#> [1110,]   0.850817401
#> [1111,]  -0.994135108
#> [1112,]   1.288115171
#> [1113,]   7.377235127
#> [1114,]  -2.768430649
#> [1115,]  -2.424376230
#> [1116,]  -4.011386249
#> [1117,]  -6.849839566
#> [1118,]   6.373952549
#> [1119,]  -5.665858225
#> [1120,]   2.515323650
#> [1121,]  -2.995677811
#> [1122,]   8.510559078
#> [1123,]   2.641073027
#> [1124,]  -1.425819542
#> [1125,]  -8.918973602
#> [1126,] -12.324978180
#> [1127,]  -9.825880250
#> [1128,]  -6.348830160
#> [1129,]   0.916123205
#> [1130,]   1.741169840
#> [1131,]  -4.100741636
#> [1132,]   6.281749765
#> [1133,]   6.149258364
#> [1134,]  -2.592590150
#> [1135,]   6.240989426
#> [1136,]   0.482291007
#> [1137,]  -4.100741636
#> [1138,]   0.552560940
#> [1139,]   1.574202484
#> [1140,]  11.489822487
#> [1141,]  10.481569351
#> [1142,]   1.648717122
#> [1143,]  -4.517348165
#> [1144,]   1.713489217
#> [1145,]  -2.093306115
#> [1146,]  -6.499273722
#> [1147,]  -1.760093058
#> [1148,]  -4.202034828
#> [1149,]  -3.823788360
#> [1150,]   0.805941935
#> [1151,]  -4.618460942
#> [1152,]  -4.425977930
#> [1153,]  18.877595955
#> [1154,]  -3.183593866
#> [1155,]  25.917152393
#> [1156,]  -3.509010574
#> [1157,]   3.648378319
#> [1158,]  -3.426158344
#> [1159,]  -3.040802345
#> [1160,]  -3.350561222
#> [1161,]  -2.185217592
#> [1162,]  -5.433208435
#> [1163,]   2.407415541
#> [1164,]  -3.851102464
#> [1165,]   0.731388937
#> [1166,]  -0.446897818
#> [1167,]  -5.489959272
#> [1168,]  -2.450922050
#> [1169,]  -4.147625393
#> [1170,]  -8.213323933
#> [1171,]  -0.701282878
#> [1172,]  -5.238432360
#> [1173,]   4.592096513
#> [1174,]  -1.984135108
#> [1175,]  -0.759371402
#> [1176,]  -1.851102464
#> [1177,]  -3.355527584
#> [1178,]  -3.569551816
#> [1179,]   2.074721700
#> [1180,]  -3.076806923
#> [1181,]   0.565111338
#> [1182,]  -2.907023442
#> [1183,]  -2.967739286
#> [1184,]  -1.531879766
#> [1185,]   2.840545864
#> [1186,]  -0.425977930
#> [1187,]   3.566220030
#> [1188,]  -0.529214496
#> [1189,]  -4.768233900
#> [1190,]   2.409499486
#> [1191,]  -1.818325023
#> [1192,]  -2.363112477
#> [1193,]  -5.586626509
#> [1194,]  -4.451313172
#> [1195,]  -4.934321212
#> [1196,]   2.424782408
#> [1197,]  -2.268069821
#> [1198,]   6.812832249
#> [1199,]  -8.016626509
#> [1200,]  -4.091501975
#> [1201,]  -3.693516823
#> [1202,]  -5.999814964
#> [1203,]  -9.000214474
#> [1204,]  -3.326158344
#> [1205,]   0.880095745
#> [1206,]   2.880456573
#> [1207,]   0.880095745
#> [1208,]   0.118415519
#> [1209,]   1.944503459
#> [1210,]  -3.618821770
#> [1211,]  -2.906843028
#> [1212,] -14.150275183
#> [1213,]  -0.991884829
#> [1214,]   0.740087356
#> [1215,] -10.296279761
#> [1216,]  -2.851463292
#> [1217,]   1.971990887
#> [1218,] -14.318325023
#> [1219,]   0.895549210
#> [1220,]  -4.354805928
#> [1221,]  -1.176099347
#> [1222,]   5.303276905
#> [1223,]  12.130817401
#> [1224,]   4.478046301
#> [1225,]  -4.426880001
#> [1226,]   5.113939336
#> [1227,]  -6.750019980
#> [1228,]  -3.833208435
#> [1229,]   2.061169840
#> [1230,]  -2.434856764
#> [1231,]  -1.858560227
#> [1232,]   1.843178997
#> [1233,]  -2.758703669
#> [1234,]  -0.258469332
#> [1235,]  -8.755713916
#> [1236,]   0.644833244
#> [1237,]   4.761567640
#> [1238,]  -4.476024559
#> [1239,]  -0.915121633
#> [1240,]  -1.851080439
#> [1241,]  -1.651824120
#> [1242,]  -4.000514594
#> [1243,]  -2.806579025
#> [1244,]   2.649077950
#> [1245,]   2.144652830
#> [1246,]  -0.152065243
#> [1247,]  -6.590802345
#> [1248,]  -3.426880001
#> [1249,]  -2.365639128
#> [1250,]   0.814962822
#> [1251,]  -1.665858225
#> [1252,]   4.074743726
#> [1253,]  -1.935398006
#> [1254,]  -9.424715032
#> [1255,]   9.627595955
#> [1256,]  -2.851102464
#> [1257,] -14.737542659
#> [1258,]  -5.819583714
#> [1259,]  -0.759551816
#> [1260,]  18.973576786
#> [1261,]  -1.768430649
#> [1262,]  -7.002486779
#> [1263,]   3.223089706
#> [1264,]  -1.685217592
#> [1265,]  -4.517889407
#> [1266,]  -7.050485784
#> [1267,]   1.073661242
#> [1268,]  -1.017348165
#> [1269,]  -3.210530808
#> [1270,]  -4.570485784
#> [1271,]  -1.092223631
#> [1272,]   0.369077950
#> [1273,]  -3.777567262
#> [1274,]   2.730062126
#> [1275,]  -1.789853419
#> [1276,]  -2.263834133
#> [1277,]   1.731539058
#> [1278,]   2.194711828
#> [1279,]   7.625323650
#> [1280,]   2.787002394
#> [1281,]  -5.721463292
#> [1282,]   4.320809012
#> [1283,]   2.210146354
#> [1284,]   1.956406134
#> [1285,]  -9.272089845
#> [1286,]  -3.426519172
#> [1287,]  -1.340936130
#> [1288,]  -5.450200394
#> [1289,]  -3.688379813