Regression Coefficients Hypothesis Test and Confidence Intervals

Regression Coefficients Hypothesis Test and Confidence Intervals

betahatinference(X, y)

Arguments

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

Value

Returns a matrix with the following columns

coef

Coefficients.

se

Standard error.

t

t-statistic.

p

p-value.

ci_0.05

Lower limit 99.99% confidence interval.

ci_0.5

Lower limit 99% confidence interval.

ci_2.5

Lower limit 95% confidence interval.

ci_97.5

Upper limit 95% confidence interval.

ci_99.5

Upper limit 99% confidence interval.

ci_99.95

Upper limit 99.99% confidence interval.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# Simple regression------------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
X <- X[, c(1, ncol(X))]
y <- jeksterslabRdatarepo::wages.matrix[["y"]]
betahatinference(X = X, y = y)

#>              coef         se         t            p   ci_0.05    ci_0.5
#> constant 4.874251 0.72698105  6.704784 3.011026e-11 2.4765930 2.9988909
#> age      0.197486 0.01834111 10.767395 6.019852e-26 0.1369951 0.1501723
#>             ci_2.5   ci_97.5   ci_99.5  ci_99.95
#> constant 3.4480531 6.3004490 6.7496112 7.2719091
#> age      0.1615042 0.2334678 0.2447997 0.2579769

# Multiple regression----------------------------------------------
X <- jeksterslabRdatarepo::wages.matrix[["X"]]
# age is removed
X <- X[, -ncol(X)]
betahatinference(X = X, y = y)

#>                  coef         se         t            p     ci_0.05     ci_0.5
#> constant   -7.1833382 1.01578786 -7.071691 2.508276e-12 -10.5335348 -9.8037324
#> gender     -3.0748755 0.36461621 -8.433184 8.939416e-17  -4.2774257 -4.0154638
#> race       -1.5653133 0.50918754 -3.074139 2.155664e-03  -3.2446781 -2.8788475
#> union       1.0959758 0.50607809  2.165626 3.052356e-02  -0.5731336 -0.2095371
#> education   1.3703010 0.06590421 20.792312 5.507605e-83   1.1529406  1.2002901
#> experience  0.1666065 0.01604756 10.382050 2.659960e-24   0.1136797  0.1252092
#>                ci_2.5    ci_97.5    ci_99.5   ci_99.95
#> constant   -9.1761258 -5.1905507 -4.5629441 -3.8331417
#> gender     -3.7901849 -2.3595660 -2.1342872 -1.8723252
#> race       -2.5642449 -0.5663817 -0.2517792  0.1140514
#> union       0.1031443  2.0888072  2.4014886  2.7650852
#> education   1.2410091  1.4995928  1.5403119  1.5876614
#> experience  0.1351242  0.1980889  0.2080039  0.2195334