Root Mean Squared Error (from \(RSS\))

Calculates the root mean squared error \(\left( RMSE \right)\) using $$ RMSE = \sqrt{\frac{1}{n} \sum_{i = 1}^{n} \left( \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}} \right)^{2}} \\ = \sqrt{\frac{1}{n} \sum_{i = 1}^{n} \left( \mathbf{y} - \mathbf{\hat{y}} \right)^{2}} \\ = \sqrt{\frac{RSS}{n}} . $$

.RMSE(MSE = NULL, X, y)

Arguments

MSE
Numeric. Mean square error.

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

Value

Returns the root mean squared error.

Details

If RSS = NULL, the RSS vector is computed using RSS() with X and y as required arguments and betahat as an optional argument. If RSS is provided, X, and y are not needed.

References

Wikipedia: Root Mean Square Deviation

See also