Adjusted R-squared \(\bar{R}^{2}\) (from \(R^2\))

Calculates the adjusted coefficient of determination $$ \bar{R}^{2} = 1 - \left(\frac{RSS / \left( n - k \right)} {TSS / \left( n - 1 \right)} \right) = 1 - \left( 1 - R^2 \right) \frac{n - 1}{n - k} . $$

.Rbar2(R2 = NULL, n, k, X, y, fromRSS = TRUE)

Arguments

R2
Numeric. Coefficient of determination \(R^2\) .

n
Integer. Sample size.

k
Integer. Number of regressors including a regressor whose value is 1 for each observation.

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

fromRSS
Logical. If TRUE, calculates the coefficient of determination from RSS. If FALSE, calculates the coefficient of determination from ESS.

Value

Returns the adjusted coefficient of determination \(\bar{R}^{2}\) .

Details

If R2 = NULL, R2 is computed using R2() with X and y as required arguments. If R2 is provided, X, and y are not needed.

References

Wikipedia: Residual Sum of Squares

Wikipedia: Explained Sum of Squares

Wikipedia: Total Sum of Squares

Wikipedia: Coefficient of Determination

See also

Author

Ivan Jacob Agaloos Pesigan