Model Assessment

Model Assessment

.model(RSS = NULL, TSS = NULL, n, k, X, y)

Arguments

RSS
Numeric. Residual sum of squares.

TSS
Numeric. Total sum of squares.

n
Integer. Sample size.

k
Integer. Number of regressors including a regressor whose value is 1 for each observation.

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

Value

Returns a vector with the following elements

RSS

Residual sum of squares.

MSE

Mean squared error.

RMSE

Root mean squared error.

R2

R-squared \(\left( R^2 \right)\).

Rbar2

Adjusted R-squared \(\left( \bar{R}^2 \right)\) .

Details

If RSS = NULL, RSS is computed using RSS() with X and y as required arguments. If RSS is provided, X, and y are not needed. If TSS = NULL, TSS is computed using TSS() with y as r equired argument. If TSS is provided, y is not needed.

References

Wikipedia: Residual Sum of Squares

Wikipedia: Explained Sum of Squares

Wikipedia: Total Sum of Squares

Wikipedia: Coefficient of Determination

See also