Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))

Standard Errors of Standardized Estimates of Regression Coefficients (Yuan and Chan (2011))

.sehatslopesprimedelta(
  slopes,
  sigma2hatepsilonhat,
  SigmaXhat,
  sigmayXhat,
  sigma2yhat,
  adjust = FALSE,
  n,
  X,
  y
)

Arguments

slopes
Numeric vector of length p or p by 1 matrix. The vector \(\boldsymbol{\hat{\beta}_{2, \cdots, k}}\) is a \(p \times 1\) vector of estimates of \(p\) unknown regression slopes.

sigma2hatepsilonhat
Numeric. Estimate of error variance.

SigmaXhat
p by p numeric matrix. Estimated \(p \times p\) variances and covariances of \({X}_{2}, {X}_{3}, \cdots, {X}_{k}\) \(\left( \boldsymbol{\hat{\Sigma}}_{\mathbf{X}} \right)\).

sigmayXhat
Numeric vector of length p or p by 1 matrix. Estimated \(p \times 1\) covariances between the regressand and the regressors \(\left( \boldsymbol{\sigma}_{\mathbf{yX}} \right)\).

sigma2yhat
Numeric. Estimated variance of the regressand \(\left( \hat{\sigma}_{y} \right)\)

adjust
Logical. Use \(n - 3\) adjustment for small samples.

n
Integer. Sample size.

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

See also

Examples

slopes <- c(-3.0748755, -1.5653133, 1.0959758, 1.3703010, 0.1666065)
SigmaXhat <- matrix(
  data = c(
    0.25018672, 0.00779108, -0.01626038, -0.04424864, -0.13217068,
    0.00779108, 0.12957466, 0.01061297, -0.08818286, -0.16427222,
    -0.016260378, 0.010612975, 0.133848763, 0.004083767, 0.658462191,
    -0.044248635, -0.088182856, 0.004083767, 7.917601877, -5.910469742,
    -0.1321707, -0.1642722, 0.6584622, -5.9104697, 136.0217584
  ),
  ncol = 5
)
sigma2hatepsilonhat <- 42.35584
sigma2yhat <- 62.35235
sigmayXhat <- c(-0.8819639, -0.3633559, 0.2953811, 10.1433433, 15.9481950)
n <- 1289
.sehatslopesprimedelta(
  slopes = slopes, sigma2hatepsilonhat = sigma2hatepsilonhat,
  SigmaXhat = SigmaXhat, sigma2yhat = sigma2yhat, sigmayXhat = sigmayXhat, n = n
)

#>      sehatslopesprime
#> [1,]       0.02282716
#> [2,]       0.02317122
#> [3,]       0.02342286
#> [4,]       0.02113537
#> [5,]       0.02330714