Residual Variance \(\hat{\sigma}_{\hat{\varepsilon}}^{2}\) (from \(\mathrm{RSS}\))

Calculates an estimate of the error variance $$ \mathbf{E} \left( \sigma^2 \right) = \hat{\sigma}_{\hat{\varepsilon}}^{2} $$ $$ \hat{\sigma}_{\hat{\varepsilon}}^{2} = \frac{1}{n - k} \sum_{i = 1}^{n} \left( \mathbf{y} - \mathbf{X} \boldsymbol{\hat{\beta}} \right)^2 \\ = \frac{\boldsymbol{\hat{\varepsilon}}^{\prime} \boldsymbol{\hat{\varepsilon}}}{n - k} \\ = \frac{\mathrm{RSS}}{n - k}$$ where \(\boldsymbol{\hat{\varepsilon}}\) is the vector of residuals, \(\mathrm{RSS}\) is the residual sum of squares, \(n\) is the sample size, and \(k\) is the number of regressors including a regressor whose value is 1 for each observation on the first column.

.sigma2hatepsilonhat(RSS = NULL, n, k, X, y)

Arguments

RSS
Numeric. Residual sum of squares.

n
Integer. Sample size.

k
Integer. Number of regressors including a regressor whose value is 1 for each observation on the first column.

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

Value

Returns the estimated residual variance \(\hat{\sigma}_{\hat{\varepsilon}}^{2}\) .

Details

If RSS = NULL, RSS is computed using RSS(). If RSS is provided, X, and y are not needed.

References

Wikipedia: Linear Regression

Wikipedia: Ordinary Least Squares

See also