Regression Standardized Slopes \(\boldsymbol{\beta}_{2, \cdots, k}^{\prime}\)

Derives the standardized slopes \(\boldsymbol{\beta}_{2, \cdots, k}^{\prime}\) of a linear regression model as a function of correlations.

.slopesprime(RX = NULL, ryX = NULL, X, y)

Arguments

RX
p by p numeric matrix. \(p \times p\) matrix of correlations between the regressor variables \(X_2, X_3, \cdots, X_k\) \(\left( \mathbf{R}_{\mathbf{X}} \right)\).

ryX
Numeric vector of length p or p by 1 matrix. \(p \times 1\) vector of correlations between the regressand variable \(y\) and the regressor variables \(X_2, X_3, \cdots, X_k\) \(\left( \mathbf{r}_{\mathbf{y}, \mathbf{X}} = \left\{ r_{y, X_2}, r_{y, X_3}, \cdots, r_{y, X_k} \right\}^{T} \right)\).

X
n by k numeric matrix. The data matrix \(\mathbf{X}\) (also known as design matrix, model matrix or regressor matrix) is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation on the first column.

y
Numeric vector of length n or n by 1 matrix. The vector \(\mathbf{y}\) is an \(n \times 1\) vector of observations on the regressand variable.

Value

Returns the standardized slopes \(\boldsymbol{\beta}_{2, \cdots, k}^{\prime}\) of a linear regression model derived from the correlation matrix.

Details

The linear regression standardized slopes are calculated using $$ \boldsymbol{\beta}_{2, \cdots, k}^{\prime} = \mathbf{R}_{\mathbf{X}}^{T} \mathbf{r}_{\mathbf{y}, \mathbf{X}} $$

where

\(\mathbf{R}_{\mathbf{X}}\) is the \(p \times p\) correlation matrix of the regressor variables \(X_2, X_3, \cdots, X_k\) and

\(\mathbf{r}_{\mathbf{y}, \mathbf{X}}\) is the \(p \times 1\) column vector of the correlations between the regressand \(y\) variable and regressor variables \(X_2, X_3, \cdots, X_k\)

See also

Author

Ivan Jacob Agaloos Pesigan