Error Variance \(\sigma_{\varepsilon}^{2}\)

Error variance \(\sigma_{\varepsilon}^{2}\) from the slopes \(\boldsymbol{\beta}_{2, \cdots, k}\) and variance-covariance matrix \(\boldsymbol{\Sigma}\).

sigma2epsilon(slopes, sigma2y, sigmayX, SigmaX)

Arguments

slopes
Numeric vector of length p or p by 1 matrix. \(p \times 1\) column vector of regression slopes \(\left( \boldsymbol{\beta}_{2, 3, \cdots, k} = \left\{ \beta_2, \beta_3, \cdots, \beta_k \right\} \right)\) .

sigma2y
Numeric. Variance of the regressand variable \(\left( \sigma_{y}^{2} \right)\).

sigmayX
Numeric vector of length p or p by 1 matrix. \(p \times 1\) vector of covariances between the regressand \(y\) variable and regressor variables \(X_2, X_3, \cdots, X_k\) \(\left( \boldsymbol{\sigma}_{\mathbf{y}, \mathbf{X}} = \left\{ \sigma_{y, X_2}, \sigma_{y, X_3}, \cdots, \sigma_{y, X_k} \right\}^{T} \right)\).

SigmaX
p by p numeric matrix. \(p \times p\) matrix of variances and covariances between regressor variables \({X}_{2}, {X}_{3}, \cdots, {X}_{k}\) \(\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)\).

See also

Examples

slopes <- c(0.2076475, 0.4510391)
sigma2y <- 1.288163
sigmayX <- c(0.4661224, 0.5771429)
SigmaX <- matrix(
  data = c(1.2934694, 0.4379592, 0.4379592, 1.0779592),
  ncol = 2
)
sigma2epsilon(slopes = slopes, sigma2y = sigma2y, sigmayX = sigmayX, SigmaX = SigmaX)

#> [1] 0.9310598