Regression Slopes \(\boldsymbol{\beta}_{2 \cdots k}\)

Derives the slopes \(\boldsymbol{\beta}_{2 \cdots k}\) of a linear regression model (\(\boldsymbol{\beta}\) minus the intercept) as a function of covariances.

slopes(SigmaX, sigmayX)

Arguments

SigmaX
p by p numeric matrix. Variances and covariances of \({X}_{2}, {X}_{3}, \cdots, {X}_{k}\) \(\left( \boldsymbol{\Sigma}_{\mathbf{X}} \right)\).

sigmayX
Vector or p by 1 matrix of covariances between the regressand and the regressors. \(\left( \boldsymbol{\sigma}_{\mathbf{yX}} \right)\).

Value

Returns the slopes \(\boldsymbol{\beta}_{2 \cdots k}\) of a linear regression model derived from the variance-covariance matrix.

Details

The linear regression slopes is given by $$ \boldsymbol{\beta}_{2 \cdots k} = \mathrm{V}_{\mathbf{X}}^{T} \mathrm{v}_{\mathbf{yX}} $$

where

\(\boldsymbol{\Sigma}_{\mathbf{X}}\) is the \(p \times p\) covariance matrix of the regressor variables and

\(\boldsymbol{\sigma}_{\mathbf{yX}}\) is the \(p \times 1\) column vector of the covariances between the regressand and the regressors.

See also